Avoid duplicate `wait_events` methods
